refactor cloud-init detection to remove import circle
